+++ packages/xc-cfg/src/builders/extrinsics/xcm/polkadotXcm.utils.ts
@@ -46,6 +46,11 @@ const isBridgeHubTransfer = (source: Parachain, destination: Parachain) => {
   );
 };
 
+export const isSnowbridgeTransfer = (assetLocation: any) => {
+  const consensus = multiloc.findGlobalConsensus(assetLocation);
+  return consensus && 'Ethereum' in consensus;
+};
+
 export const toDest = (
   version: XcmVersion,
   source: Parachain,
@@ -329,13 +334,13 @@ export const toBridgeXcmOnDest = (
  * @returns fixed location
  */
 const reanchorLocation = (assetLocation: object) => {
-  const erc20Key = multiloc.findNestedKey(assetLocation, 'key');
+  const erc20KeyObj = multiloc.findNestedKey(assetLocation, 'key');
 
-  if (erc20Key) {
+  if (erc20KeyObj) {
     return {
       parents: 0,
       interior: XcmV3Junctions.X1(
-        XcmV3Junction.AccountKey20({ key: erc20Key })
+        XcmV3Junction.AccountKey20({ key: erc20KeyObj.key })
       ),
     };
   }